Add Sub-Skeleton Creation and Skeleton Iterator
Added a new function, CreateSubSkeleton, to the Skeleton class. This function creates a sub-skeleton from the original skeleton, starting from a specified root bone and including all its descendants, stopping at the bones specified in endBoneNames.

Also added an Iterator class to the Skeleton class. This iterator allows for easy traversal of the bones in the skeleton in a depth-first manner. The iterator is initialized with the root bone ID.